Migrate to Java 24 and Gradle 8.14.1. Enhance JavaDocs.

+++ litiengine/src/main/java/de/gurkenlabs/litiengine/environment/tilemap/xml/DecimalFloatAdapter.java
@@ -3,23 +3,46 @@ package de.gurkenlabs.litiengine.environment.tilemap.xml;
 import jakarta.xml.bind.annotation.adapters.XmlAdapter;
 
 /**
- * This adapter implementation ensures that the float value is serialized like an integer when it has no digits behind
- * the decimal point.
+ * An adapter for converting between `String` and `Float` values during XML serialization and deserialization.
+ *
+ * <p>This implementation ensures that float values are serialized as integers if they have no digits
+ * after the decimal point. For example:
+ * <ul>
+ *   <li>A float value of `5.0` will be serialized as `"5"`.</li>
+ *   <li>A float value of `5.5` will remain serialized as `"5.5"`.</li>
+ * </ul>
  */
 public class DecimalFloatAdapter extends XmlAdapter<String, Float> {
 
+  /**
+   * Converts a `String` value from the XML into a `Float` object.
+   *
+   * @param v The string value to be converted.
+   * @return The parsed `Float` value.
+   * @throws Exception If the string cannot be parsed as a float.
+   */
   @Override
   public Float unmarshal(String v) throws Exception {
     return Float.parseFloat(v);
   }
 
+  /**
+   * Converts a `Float` value into a `String` for XML serialization.
+   *
+   * <p>If the float value is `null`, this method returns `null`. If the float value has no digits
+   * after the decimal point, it is serialized as an integer string. Otherwise, it is serialized as a standard float string.
+   *
+   * @param v The float value to be converted.
+   * @return The string representation of the float value.
+   * @throws Exception If an error occurs during conversion.
+   */
   @Override
   public String marshal(Float v) throws Exception {
     if (v == null) {
       return null;
     }
 
-    if (v.floatValue() % 1 == 0) {
+    if (v % 1 == 0) {
       return Integer.toString(v.intValue());
     }
